tx8m-base.inc: fix assignment of KARO_BOARD_PMIC
KARO_BOARD_PMIC is assigned with "??=" in two different places.
The empty sting assigned in linux-imx_%.bbappend wins and all boards,
that rely on the default assigned in tx8m-base.inc are left with the
empty value. Assign the default PMIC in tx8m-base.inc with "?=" to
let it take precedence over the assignment in linux-imx_%.bbappend.